==> 11-subgroups.Rmd <==
# Subgroup-based analysis {#subgroups2}

```{r, echo=FALSE, eval=TRUE, collapse=TRUE, include=TRUE, warning=FALSE, message=FALSE}
library(dplyr)
library(magrittr)
library(ggplot2)
library(LTabundR)

rerun_from_scratch <- FALSE
# If you want to confirm the package works, change to TRUE.
# If you already know it does, change to FALSE so that knitting is quick
# (it will use saved R objects instead of producing them anew)
```

False killer whales (*Pseudorca crassidens*) are rare and occur in dispersed subgroups, which complicates conventional distance sampling approaches to line-transect analysis ([Bradford et al. 2014](https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0090464)). To better estimate their abundance in Hawaiian waters, the Pacific Islands Fisheries Science Center initiated a sub-group protocol referred to as the "PC Protocol", a reference to the species' scientific name. Data collected using the PC Protocol is then analyzed using a subgroup-based analytical approach ([Bradford et al. 2014](https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0090464), [2020](https://repository.library.noaa.gov/view/noaa/24756)).  


An additional complication is that false killer whales in Hawaiian waters belong to three discrete populations -- the Main Hawaiian Islands insular population, the Northwestern Hawaiian Islands (NWHI) population, and a pelagic population -- whose ranges partially overlap, which means that population assignment cannot always be based simply on the geographic location of sightings. When geographic assignment of population is not possible, biopsy-sampled genetics or photo-identification inference, if available, is used to assign each sighting to a population *post-hoc*.  

To accommodate these special circumstances for false killer whales (and potentially other species with subgroup structure) with an appropriate balance of flexibility and efficiency, `LTabundR` includes a function named `lta_subgroup()`, whose use looks something like this: 

```{r, echo=TRUE, eval=FALSE, collapse=TRUE, message=FALSE, warning=FALSE}
lta_subgroup(df_sits,
             truncation_distance,
             ss,
             density_segments,
             density_das,
             density_sightings,
             Rg0= NULL, 
             cruz10 = NULL,
             g0_spp = NULL,
             g0_truncation = NULL,
             g0_constrain_shape = FALSE,
             g0_jackknife_fraction = 0.1,
             abundance_area = NULL,
             iterations = 5000,
             density_bootstraps = 10000,
             output_dir = NULL,
             toplot = TRUE,
             verbose = TRUE
```

Note that there are several required inputs (without any defaults) as well as many optional inputs (with defaults provided).   

We will step through each of these inputs below, using a case study in which we estimate the abundance of the pelagic false killer whale population in the Hawaiian EEZ for 2017 ([Bradford et al. 2020](https://repository.library.noaa.gov/view/noaa/24756)). 

Before proceeding, we recommend that you review the details of [how subgroup data are processed](#subgroups) within the `LTabundR` framework.  


## Inputs {-}  

### `df_sits` {-}  

This is a `data.frame` of sightings you want to use to fit the detection function model. For false killer whales in [Bradford et al. (2020)](https://repository.library.noaa.gov/view/noaa/24756), this is a combination of systematic-effort sightings prior to 2011 (made without the PC protocol) and "Phase 1" sightings from 2011 onwards (using the PC protocol). No filtering will be applied to these sightings within this function, so make sure you provide the data pre-filtered. [Bradford et al. (2020)](https://repository.library.noaa.gov/view/noaa/24756) used a single detection function for all populations of false killer whales. 

`LTabundR` has a built-in dataset for processed Central North Pacific surveys, 1986-2020, using 150-km segments. We will use that here: 

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
data("cnp_150km_1986_2020")
cruz <- cnp_150km_1986_2020
```

The code used to generate this dataset can be seen by pulling up the help documentation: `?noaa_10km_1986_2020`.  

All systematic-effort sightings from 1986-2010 are assumed to represent an initial “Phase 1” subgroup sighting (even though the PC protocol was not yet in practice), and therefore usable in detection function estimation. Here we draw those sightings from the above `cruz` object, filtering as needed (the species code for false killer whales is `"033"`), and to simplify we will select only a few key columns.   

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
sits1  <-
  cruz$cohorts$pseudorca$sightings %>%
  filter(OnEffort == TRUE,
         year < 2011,
         Lat >= 5, Lat <= 40, Lon >= -185, Lon <= -120,
         species == '033',
         ObsStd == TRUE,
         Bearing <= 90,
         mixed == FALSE) %>%
  select(DateTime, Lat, Lon, Cruise, PerpDistKm)

sits1 %>% nrow

sits1 %>% head
```

For 2011 onwards, we will use Phase 1 subgroup detections from the PC protocol, making sure that the column holding detection distances is named `PerpDistKm`:    

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
sits2  <-
  cruz$cohorts$pseudorca$subgroups$subgroups %>%
  filter(OnEffort == TRUE,
         lubridate::year(DateTime) >= 2011,
         Lat >= 5, Lat <= 40, Lon >= -185, Lon <= -120,
         ObsStd == TRUE,
         Angle <= 90,
         Species == '033',
         Phase == 1) %>%
  select(DateTime, Lat, Lon, Cruise, PerpDistKm = PerpDist)

sits2 %>% nrow

sits2 %>% head
```

To create `df_sits` for detection function fitting, we combine these datasets together:  

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
df_sits <- rbind(sits1, sits2)

df_sits %>% nrow
```

### `truncation_distance` {-}  

The truncation distance, in km, will be applied during detection function model fitting. Typically the farthest 5 - 10% of sightings are truncated, but this needs to be balanced by sample size considerations.  

Tabulate detection distances using a `LTabundR` function:  

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE, fig.height=4, fig.width=7}
dists <- 
  summarize_distances(df_sits$PerpDistKm) %>% 
  select(-total_within)

dists
```

Plot these options:  

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE, fig.height=4, fig.width=7}
ggplot(dists, aes(x=km_max_excl, y=percent_beyond)) + 
  geom_point() + geom_path() + 
  geom_hline(yintercept = 10, lty=2, color='firebrick') + 
  ylab('Percent of sightings beyond this distance') +
  xlab('Perpendicular distance (km)') + 
  scale_x_continuous(n.breaks = 10) + 
  scale_y_continuous(limits = c(0, 100), n.breaks = 10) 
```

Based on these results, we will choose a truncation distance of 4.5 km. 

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
truncation_distance <- 4.5
```

### `ss` {-}  

This is a numeric vector of subgroup school sizes. The function will find this vector's arithmetic mean and bootstrapped CV. In [Bradford et al. (2020)](https://repository.library.noaa.gov/view/noaa/24756), school size data come from all Phase 1 and Phase 2 estimates of subgroup sizes from 2011 onwards. In the processed `cruz` object, each of those estimates is the geometric mean of repeat estimates from separate observers.  

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
ss  <-
  cruz$cohort$all$subgroups$subgroups %>%
  filter(lubridate::year(DateTime) >= 2011,
         Lat >= 5, Lat <= 40, Lon >= -185, Lon <= -120,
         GSBest_geom_valid == TRUE, # only use valid "Best" estimates
         Species == '033') %>%
  pull(GSBest_geom)
  
ss %>% length

ss
```

### `Rg0` {-}  

This is a `data.frame` with estimates of Relative *g(0)* and its CV at each Beaufort sea state. If this input is left `NULL`, then these estimates will be produced by the function using the subsequent `g0_` inputs below. If this input is *not* supplied and any of the subsequent `g0_` inputs are missing, then *g(0)* will be assumed to be 1.0 with CV of 0.0.  

When you do supply this `Rg0` input, the `data.frame` has three required columns: `bft` (Beaufort sea state, numbers between 0 and 6), `Rg0` (*Rg(0)* estimates for each Beaufort state), `Rg0_CV` (the CV of the *Rg(0)* estimate in each Beaufort state). Other columns are allowed but will be ignored. 

Here is an example of a valid `Rg0` input based on the values reported for false killer whales in [Bradford et al. (2020)](https://www.fisheries.noaa.gov/inport/item/59592).  (These numbers are also available in the built-in dataset `barlow_2015`.). 

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
Rg0 <- data.frame(bft = 0:6,
                  Rg0 = c(1, 1, 0.72, 0.51, 0.37, 0.26, 0.19),
                  Rg0_CV = c(0, 0, 0.11, 0.22, 0.34, 0.46, 0.59))

Rg0
```

Again: if you supply this input, then the following `g0_` inputs will be ignored. 

### `cruz10` {-}  

This is a processed `cruz` object with short segment lengths, ideally 10 km or less (hence the 10 in the input name). This `cruz` object will be used to estimate *Rg(0)*, i.e., the relative trackline detection probability (see its chapter), using the following `g0_` inputs. `LTabundR` comes with a built-in dataset we can use for this purpose:  

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
data("noaa_10km_1986_2020")
cruz10 <- noaa_10km_1986_2020
```

The code used to generate this dataset can be seen by pulling up the help documentation: `?noaa_10km_1986_2020`.  


### `g0_spp` {-}  

This and the following `g0_` inputs will be used to model Relative *g(0)* estimates and their CV in various Beaufort sea states. If the previous input, `Rg0` is provided, then these `g0_` inputs will be ignored, and no *Rg(0)* modeling will take place. Furthermore, if any of these `g0_` inputs are not provided, *Rg(0)* will be coerced to 1.0 with a CV of 0.0 for all sea states.  

This input, `g0_spp`, is a character vector of species code(s) to use to estimate *Rg(0)*. In most cases this will be a single species, e.g., `"033"` for false killer whales.  

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
g0_spp <- '033'
```

### `g0_truncation` {-}  

The truncation distance to use when estimating *Rg(0)*. In [Bradford et al. (2020)](https://repository.library.noaa.gov/view/noaa/24756) this is 5.5 km.  

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
g0_truncation <- 5.5
```

### `g0_constrain_shape` {-}  

Some *Rg(0)* curves will not decline monotonically due to sample size issues at low Bft (0-2) or high Bft (5-6) states. To coerce monotonic decline, set this input to `TRUE`, and the function will use a shape-constrained GAM (`scam()` from package `scam`) instead of a classic `mgcv::gam()`.

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
g0_constrain_shape = FALSE
```

### `g0_jackknife_fraction` {-}  

The proportion of data to leave out within each jackknife permutation. The default is `0.1` (i.e., 10% of the data, yielding 10 jackknife loops), after [Barlow (2015)](https://onlinelibrary.wiley.com/doi/abs/10.1111/mms.12205).  

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
g0_jackknife_fraction = 0.1
```

### `density_segments` {-}  

The survey segments to be used in density/abundance estimation. For example, [Bradford et al. (2020)](https://repository.library.noaa.gov/view/noaa/24756) used 150-km segments to estimate false killer whale density in the Hawaiian EEZ in 2017. For this we can use the 1986-2020 dataset we loaded above. Note that no filtering will be applied to these segments by the `lta_subgroup()` function, so we need to filter them ourselves first: we want only systematic segments for the Hawaiian EEZ in 2017 (specifically, just cruises 1705 and 1706).  

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
cruzi <- filter_cruz(cruz = cruz,
                      analysis_only = TRUE,
                      years = 2017,
                      cruises = c(1705, 1706),
                      regions = 'HI_EEZ',
                      bft_range = 0:6,
                      eff_types = 'S',
                      on_off = TRUE)
```

From this filtered cruz object, we will isolate the segments data:  

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
density_segments <- cruzi$cohorts$all$segments
```

Since we do not want to stratify our analysis by smaller geostrata, such as the Main Hawaiian Islands, we will go ahead and coerce all stratum assignments to the Hawaiian EEZ geostratum:  

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
density_segments$stratum <- 'HI_EEZ'
```

### `density_das` {-}  

This is the complete survey data corresponding to the above segments. These data will be used to determine the proportion of survey effort occurring in each Beaufort sea state that is needed to compute weighted estimates of *g(0)* from the *Rg(0)* values. 

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
density_das <- cruzi$cohorts$all$das
```

### `density_sightings` {-}  

These are the encounters to use in density/abundance estimation. In [Bradford et al. (2020)](https://repository.library.noaa.gov/view/noaa/24756), these were the Phase 1 detections of false killer whale subgroups within the population-region-year of interest, e.g., Northwestern Hawaiian Island population sightings within the Hawaiian EEZ in 2017. No filtering is applied to these sightings within `lta_subgroups()`, so make sure only the sightings you wish to use are included and nothing more.  

In this example, since we do not have population information on hand, we will not filter detections to a specific population. Instead, we will estimate abundance of all false killer whales within the Hawaiian EEZ: 

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
density_sightings  <-
  cruz$cohorts$all$subgroups$subgroups %>%
  filter(lubridate::year(DateTime) == 2017,
         EffType == 'S',
         OnEffort == TRUE,
         PerpDist <= truncation_distance,
         Angle <= 90,
         ObsStd == TRUE,
         Species == '033',
         Phase == 1)

density_sightings %>% nrow

density_sightings %>% head
```

As above, let's make sure the geostratum assignments for these sightings are simple: 

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
density_sightings$stratum <- 'HI_EEZ'
```

### `abundance_area` {-}  

This is the area, in square km, of the region of interest. The density estimate will be scaled by this area.  

We have two options for finding this area. The first is to draw the area from our `cohort$strata` slot:   

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
cruz$strata$area[cruz$strata$stratum == 'HI_EEZ']
```

The second is to calculate it ourselves using the `LTabundR` function `strata_area()`. This second option will be useful if your study area is a complicated combination/substraction of multiple geostrata.  

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE, fig.height=7, fig.width=7}
data(strata_cnp)

abundance_area <- 
  strata_area(strata_all = strata_cnp,
              strata_keep = 'HI_EEZ')$km2

abundance_area
```

### Remaining inputs {-}  

**`iterations`**: Number of iterations to use in the various CV bootstrapping procedures occurring throughout this function, specifically: effective strip width CV estimation, subgroup size CV estimation, weighted *g(0)* CV estimation, and encounter rate estimation.  

**`output_dir`**: The path in which results RData files should be stored. If left "", the current working directory will be used.  

**`toplot`**: A Boolean, with default FALSE, indicating whether to plot various aspects of the analysis.  

**`verbose`**: A Boolean, with default TRUE, indicating whether to print status updates to the Console.  

**`density_bootstraps`**: Number of bootstrap iterations to use for the CV estimate of density and abundance specifically. This input allows this final step to use a different (typically larger) iteration size than the iterations input above.   


Here we call the `lta_subgroup()` function without re-modeling the Relative *g(0)* parameters, with only 200 iterations: 

```{r, echo=TRUE, eval=rerun_from_scratch, collapse=TRUE, message=FALSE, warning=FALSE}
results_subgroup <- 
  lta_subgroup(df_sits = df_sits,
               truncation_distance = truncation_distance,
               ss = ss,
               density_segments = density_segments,
               density_das = density_das,
               density_sightings = density_sightings,
               Rg0 = Rg0,
               abundance_area = abundance_area,
               iterations = 200,
               output_dir = 'subgroup/',
               toplot = TRUE,
               verbose = TRUE,
               density_bootstraps = 10000)
```


```{r, echo=TRUE, eval=rerun_from_scratch, collapse=TRUE, message=FALSE, warning=FALSE}
save(results_subgroup, file='results_subgroup.RData')
```

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
load('results_subgroup.RData')
```


## Outputs {-}  

The function returns a `list` with many slots, including estimates of density and abundance – along with estimates of intermediate parameters – with a CV derived from a bootstrapping routine. To demonstrate this output, we will use results from a call with only 10 bootstrap iterations.  

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
results_subgroup %>% names
```

Most of these slots hold best-estimates of parameters or sample size details:  

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
results_subgroup[c(1:17, 21:22)]
```

The `g0_details` slot includes the results from the `g0_model()` and `g0_weighted()` functions called internally by `lta_subgroup()`. See those functions' documentation pages for details.  

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
results_subgroup$g0_details %>% names
```

The `df` slot includes details of the detection function fit. See the documentation for `df_plot()` for details.  

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
results_subgroup$df %>% names
```

The `bootstraps` slot has the bootstrapped values for various parameters, in case they are useful for troubleshooting, subsequent analyses, and/or plotting:  

```{r, echo=TRUE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE}
results_subgroup$bootstraps %>% names
```

Some examples:  

```{r, echo=FALSE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE, fig.height=4, fig.width=7}
results_subgroup$bootstraps$g0 %>% hist(main='g(0) bootstraps')
```

```{r, echo=FALSE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE, fig.height=4, fig.width=7}
results_subgroup$bootstraps$D %>% hist(main='Density bootstraps')
```

```{r, echo=FALSE, eval=TRUE, collapse=TRUE, message=FALSE, warning=FALSE, fig.height=4, fig.width=7}
results_subgroup$bootstraps$N %>% hist(main='Abundance bootstraps')
```

## Behind the scenes {-}  

This function performs the following operations:

1. Fits a detection function to `df_sits` without covariates, using the `LTabundR` function `df_fit()`, in order to estimate the effective strip width (ESW).  

2. Conducts bootstrap re-sampling of the detection function fitting routine in order to estimate the CV of ESW.  

3. Estimates the arithmetic mean subgroup size based on the `ss` input.  

4. Creates a bootstrap-resampled distribution of subgroup sizes, with which CV is estimated.  

5. *Optional:* models the *Relative g(0)* in different survey conditions using the `LTabundR` function `g0_model()`. This function also estimates the CV of the *Rg(0)* estimate in each Beaufort sea state using jackknife resampling.  

6. Estimates the encounter rate (subgroup detections / trackline surveyed).  

7. Creates a bootstrap-resampled distribution of encounter rate estimates.  

8. Calculates an empirical weighted *g(0)* estimate according to the proportion of effort occurring in each Beaufort sea state, then uses an automated parameter MCMC optimization routine (see details in `LTabundR` function `g0_weighted()`) to estimate the CV of the weighted *g(0)* estimate.  

9. Creates a bootstrap-resampled distribution of the weighted *g(0)* estimate.  

10. Estimates density using the point estimates of effective strip width, subgroup size, *g(0)*, and the encounter rate.  

11. Estimates abundance by scaling the density estimate by the provided `abundance_area` input.  

12. Creates a bootstrap-resampled distribution of the density estimate by iteratively drawing values from the resampled distributions of the constituent parameters of the density equation.  

13. Creates a bootstrap-resampled distribution of the abundance estimate by scaling the density distribution by the `abundance_area` input.  


Note that this approach could theoretically be used for other species that occur in subgroups.
