==> 13-practicalities.Rmd <==
# Practicalities {#practicalities}

*Under construction!*  

### Reproducibility & project management {-} 

One of the features of the `ABUND` framework we need to retain is that each analysis is self-contained and reproducible. The folder for an analysis must have all the files and data needed for someone else to be able to replicate it. 

In the `R` package framework we are developing, the contents of a project folder will be straightforward (once the code base is bundled into a package).  


*Contents of a project folder:*  

- `DAS` file(s) with survey data  
- Stratum and study area polygon(s), as `csv`(s) (optional)  
- Group size calibration files (optional)  
- An `analysis.R` file (or perhaps a `.Rmd`), containing an adaptation of the code provided in the template above. This script will contain project-specific settings (e.g., the `load_settings()` call), which allow for the script to be reproducible.
